move AST memory metering tests to parser
